prod

+++ src/views/config/step/editEntrepot.vue
@@ -22,41 +22,60 @@
     	<span class="label">
     	Lieux de stockage
     	</span>
-
-    	<s-button icon="plus" theme="is-primary" label="Ajouter un lieu de stockage"  @onclick="addStock"></s-button>&nbsp;
-    	<s-button icon="info" theme="is-primary" label=""  @onclick="infoEntrepot=true"></s-button>
-    	<div class="notification is-warning" v-show="infoEntrepot">
+		<div style="padding-bottom: 5px">
+			<s-button icon="plus" theme="is-primary is-small" label="Ajouter"  @onclick="addStock"></s-button>&nbsp;
+			<s-button icon="info" theme="is-primary is-small" label=""  @onclick="infoEntrepot=true"></s-button>
+		</div>
+    	<div class="notification is-info" v-show="infoEntrepot">
 			<button class="delete" v-on:click="infoEntrepot=false"></button>
-			Un lieu de stockage est l'emplacement où vous allez sotcker les articles de vos clients.
+			Un lieu de stockage est l'emplacement où vous allez stocker les véhicules de vos clients.
 			 <br/> 
 			 Vous devez définir une capacité de stockage
 		</div>
-		<br/><br/>
+		
     	<div class="listStock">
-    		<div  v-for="stock in entrepot.stocks" class="itemList" >
-    			<input type="text" class="input"  style="width:150px" v-model="stock.nom" placeholder="Nom du lieu de stockage"/>
-    			&nbsp;&nbsp;<input type="number" class="input is-rounded"  style="width:100px" v-model="stock.capacite" placeholder="Capacite" />
+			 <s-draggable :refresh="refreshDrag" draggable=".itemDraggable" :datas="dataCollectionArray" @sortData="sort">
+				<div  v-for="stock in dataCollectionArray" class="itemDraggable" >
+				<div style="padding-top:5px;display:inline-block;"> <i class="fas fa-arrows-alt-v"/></div>&nbsp;&nbsp;
+				<input type="text" class="input is-rounded"  style="width:200px" v-model="stock.object.nom" placeholder="Nom du lieu de stockage"/>
+    			&nbsp;&nbsp;<input type="number" class="input is-rounded"  style="width:100px" v-model="stock.object.capacite" placeholder="Capacite" />
     			&nbsp;&nbsp;
-    			<s-button  theme="is-danger" icon="trash" label=""  @onclick="deleteStock(stock.id)"></s-button>
-    			<br/><br/>
-
-
-    		</div>
+    			<s-button  theme="is-danger" icon="trash" label=""  @onclick="deleteStock(stock.object.id)"></s-button>
+    			
+				 </div>
+			 </s-draggable>
     	</div>
+		<div class="notification is-danger" v-show="errorEntrepot.nbstock">
+				<button class="delete" v-on:click="errorEntrepot.nbstock=false"></button>
+				Il faut au moins un lieu de stockage !
+		</div>
     	<div class="notification is-danger" v-show="errorEntrepot.stock">
 				<button class="delete" v-on:click="errorEntrepot.stock=false"></button>
-				 Le nom d'un lieu de stockage est obligatoire et la capacité aussi !
+				 Le nom d'un lieu de stockage est obligatoire et sa capacité doit être supérieur à 0 !
 		</div>
+		<div class="notification is-danger" v-show="errorDeleteEntrepot">
+				<button class="delete" v-on:click="errorDeleteEntrepot=false"></button>
+				 Il y a des véhicules dans cet entrepot, vous ne pouvez pas le supprimer !<br/>
+				 Merci de déplacer les véhicules dans d'autres entrepots
+		</div>
+		<div class="notification is-danger" v-show="errorDeleteStock">
+				<button class="delete" v-on:click="errorDeleteStock=false"></button>
+				 Il y a des véhicules dans ce lieu de stockage, vous ne pouvez pas le supprimer !<br/>
+				 Merci de déplacer les véhicules vers un autre lieu de stockage
+		</div>
+    	<br/><br/> 
 		 </section>
 		<footer class="modal-card-foot">
-		  <div class="content ">
+		 
+			   <div  style="width:100%">
+				  <s-button  theme="is-danger is-small"  icon="trash" label=""   @onclick="deleteEntepot"></s-button>
+				  <div style="float:right">
+				  <s-button  theme="is-warning  is-small"  icon="ban" label="Annuler"   @onclick="$emit('back')">></s-button>
+				  <s-button  theme="is-success  is-small" icon="save" label="Enregistrer"  @onclick="save"></s-button>
+				  </div>
+				  </div>
+		   
 		  
-		    <s-button theme="is-success" icon="save" label="Enregistrer"  @onclick="save"></s-button>
-			&nbsp;&nbsp;
-				<s-button theme="is-warning" icon="ban" label="Annuler"  @onclick="$emit('back')"></s-button>
-		    &nbsp;&nbsp;
-			<s-button theme="is-danger" icon="trash" label="Supprimer"  @onclick="deleteEntepot" ></s-button>
-		  </div>
 		</footer>
     	
 
@@ -80,19 +99,46 @@ export default {
 			infoEntrepot: false,
 			errorEntrepot: {
 				nom: false,
-				stock: false
-			}
+				stock: false,
+				nbstock: false
+			},
+			dataCollectionArray: [],
+			refreshDrag: false,
+			errorDeleteEntrepot: false,
+			errorDeleteStock: false
+		}
+	},
+	watch: {
+		entrepot : function(val) {
+			this.dataCollectionArray = 	this.$orderJson(this.entrepot.stocks);
 		}
 	},
 	methods: {
+		 sort(myArray) {
+				this.$updateRang(myArray, this.dataCollectionArray);
+				this.refreshDrag = !this.refreshDrag;
+		},
 		deleteEntepot() {
-			entrepot_api.api.delete(this.agence, this.entrepot, ()=>{
-				this.$emit("back");
-			});
+			var nbArticles = 0;
+			for (var key in this.entrepot.stocks) {
+				nbArticles+=Object.keys(this.entrepot.stocks[key].articles).length;
+			}
+			if (nbArticles == 0)
+				entrepot_api.api.delete(this.agence, this.entrepot, ()=>{
+					this.$emit("back");
+				});
+			else
+				this.errorDeleteEntrepot = true;
 		},
 		deleteStock(idStock) {
-			delete this.entrepot.stocks[idStock];
-			this.$emit("refresh");
+			
+			if (Object.keys(this.entrepot.stocks[idStock].articles).length == 0) {
+				delete this.entrepot.stocks[idStock];
+				this.dataCollectionArray = 	this.$orderJson(this.entrepot.stocks);
+				this.$emit("refresh");
+			}
+			else
+				this.errorDeleteStock = true;
 		},
 		save() {
 			this.errorEntrepot.nom = this.entrepot.nom == "";
@@ -101,7 +147,9 @@ export default {
 				this.errorEntrepot.stock = this.entrepot.stocks[key].nom == "" || this.errorEntrepot.stock;
 				this.errorEntrepot.stock = this.entrepot.stocks[key].capacite == "" || this.errorEntrepot.stock;
 			}
-			var error = this.errorEntrepot.nom || this.errorEntrepot.stock;
+			this.errorEntrepot.nbstock = Object.keys(this.entrepot.stocks).length == 0
+			
+			var error = this.errorEntrepot.nom || this.errorEntrepot.stock || this.errorEntrepot.nbstock;
 
 			if (error == false) {
 				entrepot_api.api.save (this.agence, this.entrepot, (entrepot)=>{
@@ -115,6 +163,7 @@ export default {
 			var newStock = JSON.parse(JSON.stringify(entrepot_api.api.json_stock));
 			newStock.id=this.$uuid();				
 			this.entrepot.stocks[newStock.id] = newStock;
+			this.dataCollectionArray = 	this.$orderJson(this.entrepot.stocks);
 			this.$emit("refresh");
 		}
 		
@@ -133,4 +182,14 @@ export default {
 	overflow-y: auto;
 	height: 250px;
 }
+.itemDraggable {
+		border: thin solid rgb(187, 184, 184);
+		margin-top: 5px;
+		margin-right: 5px;
+		padding: 5px;
+	}
+  .itemDraggable:hover {
+		cursor: grab
+	}
+  
 </style>
