+++ src/views/config/step/listCategories.vue
@@ -4,16 +4,15 @@
 		
 		<s-button  theme="is-primary" label="Ajouter une categorie" icon="plus" @onclick="addCategorie"></s-button><br/><br/>
 		<s-draggable :refresh="refreshDrag" draggable=".itemDraggable" :datas="dataCollectionArray" @sortData="sort">
-		    <div v-for="categorie in dataCollectionArray" :key="categorie.id" class="itemDraggable">
-		        <s-button theme="is-primary" label="" icon="cog" @onclick="editCategorie(categorie.id)"/>
+		    <div v-for="categorie in dataCollectionArray" :key="categorie.id" class="itemDraggable" >
+		        <i class="fas fa-arrows-alt-v"/>
+				
 				<div class="itemCategory">
 					<i :class="categorie.object.icon"/>
 					&nbsp;{{categorie.object.nom}}	
 				</div>
-				<s-button theme="" label="" icon="arrows-alt" @onclick=""/>
-				 <br/><br/>
+				<s-button theme="is-primary" label="" icon="cog" @onclick="editCategorie(categorie.id)"/>
 		    </div>
-		    
 		</s-draggable>
 		<div class="notification is-danger" v-show="error">
 		  <button class="delete" v-on:click="error=false"></button>
@@ -62,65 +61,156 @@
 				  &nbsp;
 				  <s-button theme="is-primary"  label="" icon="info" @onclick="infoType=true"/>
 				  <br/><br/>
-				  <div class="notification is-warning" v-show="infoType">
+				  <div class="notification is-info" v-show="infoType">
 					  <button class="delete" v-on:click="infoType=false"></button>
-					  Un type vous permet de définir le prix par défaut par une taille ou pour autre critère ...<br/>
-					  Par exemple, une caravane de type "Petite" a un prix de 150 €.<br/>
-					  <b>Attention : Si le champ figé est coché, le prix ne pourra pas être spécifique pour un article donné.
-					  Si il n'est pas coché vous pourrez redéfinir un prix pour un article et pour un client donné</b>
+					  Un type vous permet de définir le prix et l'espace pris par défaut pour une taille ou pour un autre critère d'une catégorie de véhicule...<br/>
+					  Par exemple, une caravane de type "Petite" a un prix de 150 € prend une place de 0,5.<br/>
+					  Seul le prix pourra être redéfini par client.
 				 </div>
 				 <div class="notification is-danger" v-show="errorCategorie.type">
 					  <button class="delete" v-on:click="errorCategorie.type=false"></button>
 					  Le nom et le prix doivent être saisis.
 					  Le prix doit être un montant !
-				 </div>
-				 <table class="arrayType">
-				 		<tr>
-				 			<th>Nom</th>
-				 			<th>Prix</th>
-				 			<th>Figé</th>
-				 			<th>Place </th>
-				 			<th></th>
-				 		</tr>
-				  		<tr v-for="type in categorie.types">
-				  			<td class="nom">
-				  			 <input type="text"  class="input is-rounded" style="width:100%" v-model="type.nom" placeholder="Mettre le nom du type" />
-				  			 </td>
-				  			 <td class="prix">
-				  			 <span class="control has-icons-left has-icons-right">
-								<input type="number" style="width:100%" class="input is-rounded" v-model="type.prix"  placeholder="Prix par défaut"/>
+				</div>
+				<div class="head">
+					<div class="move"></div>
+					<div class="nom">Nom</div>
+					<div class="prix">Prix à l'année</div>
+					<div class="place">Espace pris</div>
+					
+				</div>
+				 <s-draggable :refresh="refreshDragType" draggable=".itemDraggableType" 
+				 			:datas="dataCollectionArrayTypes" @sortData="sortType">
+					<div v-for="type in dataCollectionArrayTypes"  :key="type.id" class="itemDraggableType">
+						<div class="move">
+							<i class="fas fa-arrows-alt-v"/>
+						</div>
+						<div class="nom">
+						<input type="text"  class="input is-rounded" v-model="type.object.nom" placeholder="Mettre le nom du type" />
+						</div>
+						<div class="prix">
+						<span class="control has-icons-left">
+								<input type="number"  class="input is-rounded" v-model="type.object.prix"  placeholder="Prix par défaut"/>
 								<span class="icon is-small is-left">
 								<i class="fas fa-euro-sign"></i>
-							    </span>	&nbsp;	
-							    				    
-							</span> 
-				  			</td>
-				  			<td class="fige">
-				  				<input type="checkbox" v-model="type.fige"/>	
-				  			</td>
-							<td colspan="2" class="espace">
-								<input type="number" placeholder="Nb place" v-model="type.espace" class="input is-rounded"style="width:100%" />							
-							</td>
-							<td class="trash">
-								<s-button icon="trash" label="" theme="is-small is-danger" @onclick="deleteType(type.id)"/>
-							</td>
-				  		</tr>
-				  </table>
+							    </span>	&nbsp;    
+						</span> 
+						</div>
+						<div class="place">
+						<input type="number" placeholder="Nb place" v-model="type.object.espace" class="input is-rounded"style="width:100%" />
+						</div>
+						<div class="trash">
+						<s-button icon="trash" label="" theme="is-small is-danger" @onclick="deleteType(type.object.id)"/>
+						</div>
+					</div>
+				</s-draggable>
+				
 				  <div class="notification is-danger" v-show="errorCategorie.nbType">
 					  <button class="delete" v-on:click="errorCategorie.nbType=false"></button>
 					  Il faut au moins un type !
 				</div>
 				  
+				  <div class="notification is-danger" v-show="errorNbArticles">
+					  <button class="delete" v-on:click="errorNbArticles=false"></button>
+					  Il existe des véhicules pour ce type de catégorie
+				</div>
 				</section>
 				<footer class="modal-card-foot">
-				  <s-button  theme="is-success" icon="save" label="Enregistrer"  @onclick="saveCategorie"></s-button>
-				  <s-button  theme="is-warning"  icon="ban" label="Annuler"  @onclick="modalEditCategorie='modal'"></s-button>
-				  <s-button  theme="is-danger"  icon="trash" label="Supprimer"  @onclick="deleteCategorie"></s-button>
+				  <div  style="width:100%">
+				  <s-button  theme="is-danger is-small"  icon="trash" label=""  @onclick="deleteCategorie"></s-button>
+				  <div style="float:right">
+				  <s-button  theme="is-warning  is-small"  icon="ban" label="Annuler"  @onclick="modalEditCategorie='modal'"></s-button>
+				  <s-button  theme="is-success  is-small" icon="save" label="Enregistrer"  @onclick="saveCategorie"></s-button>
+				  </div>
+				  </div>
 				</footer>
 			</div>
 		</div>
 	</div>
 </template>
+<style scoped>
+	.arrayType {
+		width: 100%;
+		
+	}
+	.arrayType tr th {
+		
+		text-align: center;
+	}
+	div.nom {
+		width: 25%;
+		text-align: center;
+		
+	}
+	div.prix {
+		width: 33%;
+		text-align: center;
+		
+	}
+	div.place {
+		width: 22%;
+		text-align: center;
+		
+	}
+	
+	div.trash {
+		width: 5%;
+		margin-top: 5px;
+		margin-left: 50px;
+	}
+	div.move {
+		width: 5%;
+		padding-top: 10px;
+	}
+	div.move button:hover{
+		cursor:move;
+	}
+	.itemCategory {
+		display: inline-block;
+		margin-top: 10px;
+		width: 150px;
+		padding-left: 15px;
+	}
+	.itemDraggable {
+		border: thin solid rgb(187, 184, 184);
+		margin-bottom: 3px;
+		margin-right: 5px;
+		padding: 5px;
+	}
+	.itemDraggableType {
+		border: thin solid rgb(187, 184, 184);
+		margin-bottom: 3px;
+		margin-right: 5px;
+		padding: 5px;
+	}
+	.itemDraggable:hover {
+		cursor: grab
+	}
+	.itemDraggableType:hover {
+		cursor: grab
+	}
+	.itemDraggable:hover {
+		cursor: grab
+	}
+	.itemDraggableType:hover {
+		cursor: grab
+	}
+	
+	.itemDraggableType div {
+		display: inline-block;
+		height: 40px;
+		vertical-align: top;
+		padding-left: 5px;
+		
+	}
+	.head div {
+		display: inline-block;
+		height: 40px;
+		vertical-align: top;
+		padding-left: 5px;
+		font-weight: bolder;
+	}
+</style>
 <script>
 	import categorie_api from '@/firebase/categorie_api'
 	import library_icons from "@/firebase/library_icon"
@@ -130,6 +220,7 @@
 		data: function() {
 			return {
 				dataCollectionArray: [],
+				dataCollectionArrayTypes: [],
 				fields: ['label', 'value', 'icon'],
 				categorie: categorie_api.api.json_categorie,
 				type: categorie_api.api.json_type,
@@ -143,7 +234,9 @@
 				},
 				infoType: false	,
 				error: false,
-				refreshDrag: false		
+				refreshDrag: false,
+				refreshDragType: false,
+				errorNbArticles: false
 			}
 		},
 		watch: {
@@ -156,6 +249,10 @@
 				this.$updateRang(myArray, this.dataCollectionArray);
 				this.refreshDrag = !this.refreshDrag;
 			},
+			sortType(myArray) {
+				this.$updateRang(myArray, this.dataCollectionArrayTypes);
+				this.refreshDragType = !this.refreshDragType;
+			},
 			setIcon(item) {
 				
 				this.categorie.icon = item.object.value;
@@ -164,10 +261,17 @@
 				var type= JSON.parse(JSON.stringify(this.type));
 				type.id = this.$uuid();
 				this.categorie.types[type.id]  = type;
+				this.dataCollectionArrayTypes = this.$orderJson( this.categorie.types);
 				this.$emit("refresh");
 			},
 			deleteType(idType) {
-				delete this.categorie.types[idType]
+				if (this.categorie.types[idType].nbArticles != 0)
+					this.errorNbArticles = true;
+				else {
+					delete this.categorie.types[idType];
+					this.errorNbArticles = false;
+					this.dataCollectionArrayTypes = this.$orderJson( this.categorie.types);
+				}
 			},
 			
 			saveCategorie() {
@@ -192,19 +296,28 @@
 				}
 			},
 			editCategorie(idCategorie) {
-				this.categorie = this.agence.categories[idCategorie]
+				this.categorie = this.agence.categories[idCategorie];
+				this.dataCollectionArrayTypes = this.$orderJson(this.agence.categories[idCategorie].types);
 				this.modalEditCategorie='modal is-active';
 			},
 			addCategorie() {				
 				this.categorie = JSON.parse(JSON.stringify(categorie_api.api.json_categorie));
 				this.categorie.id = this.$uuid();
+				this.dataCollectionArrayTypes = [];
 				this.modalEditCategorie='modal is-active';
 			},
 			deleteCategorie() {
-
-				categorie_api.api.delete(this.agence, this.categorie, ()=>{
-					this.modalEditCategorie='modal';
-				});
+				var deleteCategorie = true;
+				for (var key in this.categorie.types) {
+					if (this.categorie.types[key].nbArticles != 0)
+						deleteCategorie = false;
+				}
+				if (deleteCategorie)
+					categorie_api.api.delete(this.agence, this.categorie, ()=>{
+						this.modalEditCategorie='modal';
+					});
+				else
+					this.errorNbArticles  = true;
 			},
 			save(fct) {
 				this.error = false;
@@ -218,47 +331,7 @@
 		},
 		mounted() {
 			this.dataCollectionArray = 	this.$orderJson(this.agence.categories);
+			
 		}
 	}
 </script>
-<style scoped>
-	.arrayType {
-		width: 100%;
-		
-	}
-	.arrayType tr th {
-		
-		text-align: center;
-	}
-	.arrayType tr td.nom {
-		width: 35%;
-		text-align: center;
-
-	}
-	.arrayType tr td.prix {
-		width: 30%;
-		text-align: center;
-		
-	}
-	.arrayType tr td.place {
-		width: 25%;
-		text-align: center;
-		
-	}
-	.arrayType tr td.fige {
-		width: 5%;
-		text-align: center;
-		padding-top: 10px;
-	}
-	.arrayType tr td.trash {
-		width: 5%;
-		text-align: center;
-		
-	}
-	.itemCategory {
-		display: inline-block;
-		margin-top: 10px;
-		width: 150px;
-		padding-left: 15px;
-	}
-</style>
